app/domains/performance/application/response/festival_summary_response.py
from pydantic import BaseModel


class FestivalSummaryResponse(BaseModel):
    mt20id: str
    prfnm: str
    prfpdfrom: str
    prfpdto: str
    fcltynm: str
    poster: str
    genrenm: str
    prfstate: str
    area: str
    festival: str

    model_config = {"from_attributes": True}
