requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
aiomysql>=0.2.0
cryptography>=44.0.0
httpx>=0.27.0
apscheduler>=3.10.0
pydantic-settings>=2.0.0
beautifulsoup4>=4.12.0
